+++ arcane_docs/lib/custom.dart
@@ -3726,7 +3726,7 @@ ArcaneScreen(
 """,
       child: SizedBox(
         height: 500,
-        child: ArcaneScreen(
+        child: Screen(
             header: Bar(titleText: "Header", trailing: [
               IconButton(
                 icon: Icon(Icons.activity),
@@ -3795,7 +3795,7 @@ ArcaneScreen(
 """,
       child: SizedBox(
         height: 500,
-        child: ArcaneScreen(
+        child: Screen(
             gutter: false,
             header: Bar(titleText: "Header", trailing: [
               IconButton(