+++ lib/util/promotions.dart
@@ -27,6 +27,7 @@ export 'package:flutter/widgets.dart'
         FlexColumnWidth,
         TableCellVerticalAlignment;
 export 'package:flutter_animate/flutter_animate.dart' hide ShimmerEffect;
+export 'package:flutter_hooks/flutter_hooks.dart';
 export 'package:font_awesome_flutter/font_awesome_flutter.dart';
 export 'package:haptic_feedback/haptic_feedback.dart';
 export 'package:meta_seo/meta_seo.dart';